Leaderboard | Use quiz-leaderboard component
