+++ app/controllers/index.js
@@ -1,6 +1,7 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
+  storage: Ember.inject.service(),
 
   username: '',
   isInvalid: false,
