+++ app/initializers/storage.js
@@ -1,6 +1,5 @@
 export function initialize (application) {
   application.inject('route', 'storage', 'service:storage');
-  application.inject('controller', 'storage', 'service:storage');
 }
 
 export default {
