+++ app/templates/leaderboard.hbs
@@ -1,34 +1,1 @@
-{{ui-topbar}}
-
-<div class="Content">
-  <div class="Grid Grid--centered">
-    <div class="Grid-cell Grid-cell--10 Grid-cell--6@tablet">
-      <header class="Leaderboard-header">
-        <h2>Leaderboard</h2>
-        <div class="Tabs u-mgv--m">
-          {{#link-to "quiz" class="Tab"}}Play again{{/link-to}}
-          <span class="Tab" tabindex="1" {{action "onResetLeaderboard"}}>Reset</span>
-        </div>
-      </header>
-      <div class="List">
-        <header class="List-header">
-          <h4 class="List-title">Name</h4>
-          <strong>Score</strong>
-        </header>
-        <ul class="List-content">
-          {{#each scores as |item|}}
-            <li class="List-item">
-              <strong class="u-push">{{item.score}}</strong>
-              {{item.name}}
-            </li>
-          {{/each}}
-          {{#unless model.length}}
-            <li class="List-item">
-              No points scored yet
-            </li>
-          {{/unless}}
-        </ul>
-      </div>
-    </div>
-  </div>
-</div>
+{{quiz-leaderboard model=model onResetLeaderboard=(route-action "onResetLeaderboard")}}
